fix : 메인페이지 bookmark 오류 수정

+++ src/main/java/com/hotel/mapper/BookmarkMapper.java
@@ -10,5 +10,6 @@ public interface BookmarkMapper {
 	
 	public int deleteBookmark(@Param("miId") String miId, @Param("siId") int siId);
 	
-	public List<Long> getBookmarkList(String miId);
+	// 회원별 북마크 된 숙소번호 리스트
+	public List<Integer> getBookmarkList(String miId);
 }
